0.9.3.1 1. При выобре карандаша или инструмента выделения - на вкладке mask в окне origin картинка курсора принимает правильное значение; 2. Кнопка Repain для карандаша, теперь учитывает инверсию цветов; 3. Убрана процентная строка загрузки clog файла для большей скорости открытия. 4. Всякие мелочи.

+++ viewer/checkFile/checkfile.cpp
@@ -5,7 +5,7 @@ ListData::ListData(QString fileName)
 {
      this->checkFile(fileName);
 }
-QList<quint16> ListData::checkFile(QString fileName)
+QList<int> ListData::checkFile(QString fileName)
 {
     QFile file(fileName);
 
@@ -27,12 +27,11 @@ QList<quint16> ListData::checkFile(QString fileName)
             QStringList splitString = QString(file.readLine()).remove("\r\n").split(" ");
             int countInLine = splitString.length();
 
-            //максимальное число элемента - 65535, все что больше будет = 0, отрицательные числа будут равны: 65536 плюс отрицательное число
             foreach (QString element, splitString)
-                list.append(element.toUShort());
+                list.append(int(element.toDouble()));
 
             //если в текущей строке элементов меньше чем в максимальной строке - добавляем нули
-            while (countInLine != maxCountElement)
+            while (int(countInLine )!= maxCountElement)
             {
                 list.append(0);
                 countInLine++;
