+++ viewer/checkFile/checkfile.h
@@ -7,11 +7,11 @@ struct ListData
 {
     int row     = 0;
     int column  = 0;
-    QList<quint16>  list;
+    QList<int>  list;
 
     ListData(QString fileName);
 
-    QList<quint16> checkFile(QString fileName);
+    QList<int> checkFile(QString fileName);
 };
